Add cast to fix passing wrong size to LittleWord
(cherry picked from commit 1459f1029d876aa9879cd077f7a4389ed4d2b393)

+++ src/steamnetworkingsockets/clientlib/steamnetworkingsockets_udp.cpp
@@ -581,7 +581,7 @@ int CConnectionTransportUDPBase::SendEncryptedDataChunk( const void *pChunk, int
 			if ( usecTimeSinceSentLast <= (uint64)k_usecTimeSinceLastPacketMaxReasonable ) // Force unsigned comparison in case assert above fails
 			{
 				// Serialize it
-				out.PutUint16( LittleWord( usecTimeSinceSentLast >> k_usecTimeSinceLastPacketSerializedPrecisionShift ) );
+				out.PutUint16( LittleWord( (uint16)( usecTimeSinceSentLast >> k_usecTimeSinceLastPacketSerializedPrecisionShift ) ) );
 				out.hdr.m_unMsgFlags |= out.hdr.kFlag_TimeSincePrev;
 			}
 		}
